Eunbi

+++ sweethome/src/main/java/com/thegenesis/sweethome/community/controller/CommunityControler.java
@@ -10,10 +10,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.google.gson.Gson;
 import com.thegenesis.sweethome.common.template.Pagination;
 import com.thegenesis.sweethome.common.template.saveFile;
 import com.thegenesis.sweethome.common.vo.PageInfo;
@@ -21,6 +24,8 @@ import com.thegenesis.sweethome.common.vo.Report;
 import com.thegenesis.sweethome.community.model.service.CommunityService;
 import com.thegenesis.sweethome.community.model.vo.Community;
 import com.thegenesis.sweethome.community.model.vo.CommunityFile;
+import com.thegenesis.sweethome.community.model.vo.Reply;
+
 
 @Controller
 public class CommunityControler {
@@ -37,7 +42,7 @@ public class CommunityControler {
 	}
 	//글 상세보기
 	@RequestMapping("detail.co")
-	public ModelAndView boardDetail(ModelAndView mv, int bno) {
+	public ModelAndView boardDetail(ModelAndView mv, int bno, HttpSession session) {
 			
 		int result = communityService.increaseCount(bno);		
 			
@@ -53,8 +58,8 @@ public class CommunityControler {
 						
 		}else {
 			
-			mv.setViewName("Redirect:/");
-			//나중에 alertMsg로 바꿔주기~
+			session.setAttribute("alertMsg", "상세조회 실패");
+			mv.setViewName("community/noticeList");
 		}		
 		return mv;
 	}
@@ -239,7 +244,6 @@ public class CommunityControler {
 	}
 	
 	
-	
 	//게시글 쓰기(글작성 폼 띄워줌)
 	@RequestMapping("insertBoardView.co")
 	public ModelAndView insertBoardView(ModelAndView mv, int bType) {
@@ -318,17 +322,19 @@ public class CommunityControler {
 	@RequestMapping("update.co")
 	public String updateBoard(Community cm, CommunityFile cf, Model model, HttpSession session, MultipartFile reupfile ) {
 		
-		
 		//새로 첨부파일이 넘어온 경우
 		if(!reupfile.getOriginalFilename().equals("")) {
-			//기존 첨부파일이 있는 경우
-			//첨부 파일 삭제
-			new File(session.getServletContext().getRealPath(cf.getFilePath())).delete();
+			if(cf.getOriginName() !=null) {
+				//기존 첨부파일이 있는 경우--> 첨부 파일 삭제				
+				new File(session.getServletContext().getRealPath(cf.getFilePath())).delete();
+			}
+			
 			//새로운 첨부파일 등록
 			String changeName = saveFile.changeFileName(reupfile, session);
 			
 			cf = CommunityFile.builder().originName(reupfile.getOriginalFilename())
 						.fileNo(cf.getFileNo())
+						.boardNo(cm.getBoardNo())
 						.originName(reupfile.getOriginalFilename())
 						.changeName(changeName)
 						.filePath("resources/uploadFiles/" + changeName)
@@ -379,8 +385,7 @@ public class CommunityControler {
 	}
 	//게시글 신고
 	@RequestMapping("reportBoard.co")
-	public String reportBoard(int boardNo, int reportCate, int userNo, HttpSession session) {
-		
+	public String reportBoard(int boardNo, int reportCate, int userNo, HttpSession session) {		
 		
 		HashMap<String, String> map = new HashMap<>(); 
 		map.put("boardNo", Integer.toString(boardNo));
@@ -411,9 +416,106 @@ public class CommunityControler {
 			session.setAttribute("alertMsg", "이미 신고하셨습니다.");
 			return "redirect:detail.co?bno=" + boardNo;
 		}
-		
-		
+				
 	}
+	
+		//댓글 리스트
+		@ResponseBody
+		@RequestMapping(value="replyList.co", produces="application/json; charset=UTF-8")
+		public String ajaxSelectReplyList(int boardNo) {
+		
+			ArrayList<Reply> replyList = communityService.selectReplyList(boardNo);
+					
+			return new Gson().toJson(replyList);
+		}
+			
+		//댓글 작성
+		@ResponseBody
+		@RequestMapping(value="replyInsert.co")
+		public String ajaxInsertReply(Reply rp) {
+			
+			int result = communityService.insertReply(rp);		
+			
+			return (result > 0 ? "YY" : "NN");
+		}
+		
+		//댓글 수정
+		@RequestMapping("updateReply.co")
+		public String updateReply(Reply rp, HttpSession session) {
+			
+			int boardNo = rp.getBoardNo();
+			
+			int result = communityService.updateReply(rp);
+			
+			if(result>0) {
+				session.setAttribute("alertMsg", "댓글 수정 성공");
+				return "redirect:detail.co?bno=" + boardNo;
+			}else {
+				session.setAttribute("alertMsg", "댓글 수정 실패");
+				return "redirect:detail.co?bno=" + boardNo;
+			}
+			
+		}
+		
+		//댓글 삭제
+		@RequestMapping(value="deleteReply.co",  method = RequestMethod.GET)
+		public String deleteReply(@RequestParam(value="replyNo") int replyNo, @RequestParam(value="boardNo")int boardNo, HttpSession session) {
+			
+			int result = communityService.deleteReply(replyNo);
+			
+			if(result>0) {
+				session.setAttribute("alertMsg", "댓글 삭제 성공");
+				return "redirect:detail.co?bno=" + boardNo;
+			}else {
+				session.setAttribute("alertMsg", "댓글 삭제 실패");
+				return "redirect:detail.co?bno=" + boardNo;
+			}
+
+		}
+		//댓글 신고
+		@RequestMapping("reportReply.co")
+		public String reportReply(int boardNo, int replyNo, int reportCate, int userNo, HttpSession session) {
+			
+			System.out.println(boardNo);
+			
+			HashMap<String, String> map = new HashMap<>(); 
+			map.put("replyNo", Integer.toString(replyNo));
+			map.put("UserNo", Integer.toString(userNo));
+			
+			int check = communityService.reportCheckReply(map);//신고여부체크
+				
+			if(check == 0) {
+				Report r = new Report();
+				
+				r = Report.builder()
+						.replyNo(replyNo)
+						.reportCate(reportCate)
+						.userNo(userNo)
+						.build();
+			
+				
+				int result = communityService.reportReply(r);
+				
+				System.out.println(result);
+				
+				if(result>0) {
+					session.setAttribute("alertMsg", "신고완료");
+					return "redirect:detail.co?bno=" + boardNo;
+				}else {
+					session.setAttribute("alertMsg", "신고 실패");
+					return "redirect:detail.co?bno=" + boardNo;
+				}
+				
+			}else {
+				session.setAttribute("alertMsg", "이미 신고하셨습니다.");
+				return "redirect:detail.co?bno=" + boardNo;
+			}
+					
+		}
+		
+		
+		
+
